+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add option to warn if HTML title tag is missing or empty.
 * Add option to warn if HTML meta keywords is missing or empty.
 * Add option to warn if HTML meta description is missing or empty.
+* Add option to bypass Content-Security-Policy.
 
 ## 1.1.1 - 2022-05-06